stringArray and execArgs

+++ src/input-schema/index.ts
@@ -7,8 +7,18 @@ import { type CollapseIntersection } from '../utility-types/collapse-intersectio
 import { writeFileSync, readFileSync, existsSync, mkdirSync } from 'fs';
 import { checkIntegrity, hashConfigurationFile } from '../versioning/config-file-hash/index.ts';
 import { diffConfigurations } from '../versioning/config-diff/index.ts';
+import { arrayPropertySchema, type ArrayPropertyType } from './array-property/index.ts';
+import { execArgs, initializeExecArgs } from './exec-args.ts';
+import { yellowBG } from '../text-coloring/index.ts';
+import { exit } from 'process';
 
-const propertyTypesSchemas = [stringPropertySchema, booleanPropertySchema, integerPropertySchema, enumPropertySchema];
+const propertyTypesSchemas = [
+    stringPropertySchema,
+    booleanPropertySchema,
+    integerPropertySchema,
+    enumPropertySchema,
+    arrayPropertySchema,
+];
 const propertyTypes = z.union(propertyTypesSchemas);
 
 type AnyProperty = z.infer<typeof propertyTypes>;
@@ -23,6 +33,8 @@ type inferProperty<Input extends z.infer<(typeof propertyTypesSchemas)[number]>>
     ? IntegerPropertyInput
     : Input extends z.infer<typeof enumPropertySchema> & { enum: (infer Item)[] }
     ? Item
+    : Input extends z.infer<typeof arrayPropertySchema>
+    ? ArrayPropertyType<Input>
     : never;
 
 type inferPropertyTypesSchemas<Input extends z.infer<(typeof propertyTypesSchemas)[number]>> = Input extends {
@@ -47,6 +59,15 @@ type consistentRequired<T extends InputSchema> = T extends { properties: Record<
     ? T & { required?: R[] }
     : never;
 
+function createPathToFile(path: string) {
+    const splitPath = path.split('/');
+    for (let i = 0; i < splitPath.length - 1; i++) {
+        const folder = splitPath.slice(0, i + 1).join('/');
+        if (!existsSync(folder)) {
+            mkdirSync(folder, { recursive: true });
+        }
+    }
+}
 /**
  * @param input The input schema to be emmited
  * @returns The same input schema, to be used for input type inference
@@ -68,31 +89,33 @@ export function defineInputConfiguration<
     // `InputConfiguration` is not specific enough, so we need to add the required fields explicitly
     // using the `consistentRequired` type would end up with a union of all property keys (which breaks the inference)
 ): InputConfiguration & { required?: Requireds[] } {
-    const args = process.argv;
-    console.log(args);
+    initializeExecArgs();
     const path = '.actor/input_schema.json';
     const configExists = existsSync(path);
     const hashedInput = hashConfigurationFile(input, inputSchema);
     if (!configExists) {
         console.log('No input schema found, will write to .actor/input_schema.json');
-        const splitPath = path.split('/');
-        for (let i = 0; i < splitPath.length - 1; i++) {
-            const folder = splitPath.slice(0, i + 1).join('/');
-            if (!existsSync(folder)) {
-                mkdirSync(folder, { recursive: true });
-            }
+        createPathToFile(path);
+        if (execArgs.dryRun()) {
+            process.exit(1);
         }
         writeFileSync('.actor/input_schema.json', JSON.stringify(hashedInput, null, 4));
         return input;
     }
     const previousConfig = readFileSync('.actor/input_schema.json', 'utf-8');
     const integrity = checkIntegrity(previousConfig, inputSchema);
-    const parsedPreviousConfig = inputSchema.parse(JSON.parse(previousConfig));
+    const parsedPreviousConfig = inputSchema.safeParse(JSON.parse(previousConfig));
     if (integrity === 'Passed') {
         diffConfigurations(parsedPreviousConfig, input);
     } else {
         diffConfigurations(parsedPreviousConfig, input);
-        console.warn('Input schema changed manually, check changes');
+        if (execArgs.overwrite()) {
+            console.log('Input schema changed manually, overwriting');
+            writeFileSync('.actor/input_schema.json', JSON.stringify(hashedInput, null, 4));
+        } else {
+            console.warn(`${yellowBG('WARNING:')} Input schema changed manually, check changes`);
+            process.exit(1);
+        }
     }
     return input;
 }
